requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-restful==0.3.10
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
flask-jwt-extended==4.5.3
werkzeug==2.3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
